+++ config/laravel-chained-translator.php
@@ -18,4 +18,30 @@ return [
     | the directory does not exist yet.
     */
     'add_gitignore_to_custom_lang_directory' => true,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Group translation keys in to arrays
+    |--------------------------------------------------------------------------
+    | You can choose if the translations keys are in dotted notation or grouped
+    | using arrays.
+    |
+    | True: saved as nested arrays, f.e.
+    |   'key' => [
+    |       'detail' => 'translation',
+    |   ]
+    |
+    | False: saved as dotted keys, f.e.
+    |   'key.detail' => 'translation',
+    |
+    */
+    'group_keys_in_array' => false,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Json group name
+    |--------------------------------------------------------------------------
+    | You can customize what group is used for all json translations.
+    */
+    'json_group' => 'json-file',
 ];